requirements.txt
apache-libcloud==0.20.1
backports.ssl-match-hostname==3.5.0.1
